// src/contracts/generated/maverick-schema.d.ts
export interface components {
  schemas: {
// ...
    Event: {
      /** Format: int64 */
      doc?: number;
      /** Format: int64 */
      fields?: number;
      /** Format: int64 */
      id: number;
      name: string;
      /** Format: int64 */
      venueId: number;
      /** Format: int64 */
      venueConfigurationId: number;
      /** Format: date-time */
      eventTimeLocal: string;
      /** Format: date-time */
      endDate?: string;
      timeTbd: boolean;
      dateTbd: boolean;
      url?: string;
      parking: boolean;
      /** Format: int64 */
      parkingEventId?: number;
      /** Format: int32 */
      minimumAge?: number;
      ifNecessary: boolean;
      multiDate: boolean;
      /** Format: date-time */
      eventTimeUtc?: string;
      /** Format: int64 */
      cmsPageId?: number;
      eventPerformers?: components['schemas']['EventPerformer'][];
      venue?: components['schemas']['Venue'];
      slug?: string;
      venueConfiguration?: components['schemas']['VenueConfiguration'];
      seasonTicket: boolean;
      /** Format: date-time */
      nextOnsaleTimeLocal?: string;
      /** Format: date-time */
      nextPresaleTimeLocal?: string;
      /** Format: date-time */
      lastOnsaleTimeLocal?: string;
      /** Format: date-time */
      lastPresaleTimeLocal?: string;
      /** Format: date-time */
      firstOnSaleTimeLocal?: string;
      /** Format: int32 */
      daysToEvent?: number;
      /** Format: int32 */
      daysToOnsale?: number;
      /** Format: int32 */
      daysToPresale?: number;
      /** Format: int32 */
      daysUntilFirstOnSale?: number;
      /** @enum {string} */
      pricingType?: 'DEFAULT' | 'ALL_IN_TICKET_DETAILS' | 'ALL_IN_TICKET_PAGE';
      /** Format: int32 */
      availableTickets?: number;
      /** Format: double */
      percentInventoryAvailable?: number;
    };
// ...
    Order: {
      /** Format: int64 */
      id: number;
      /** Format: date-time */
      createdAt: string;
      status: components['schemas']['OrderStatus'];
      emailAddress: string;
      phoneNumber: string;
      deliveryMethod?: string;
      deliveryMethodDetails?: components['schemas']['DeliveryMethod'];
      serviceFee: number;
      taxAmount: number;
      deliveryFee: number;
      insuranceFee: number;
      hasInsurance: boolean;
      /** @enum {string} */
      insuranceOfferStatus?: 'NO_OFFER' | 'ACCEPTED' | 'DECLINED';
      otherFee?: number;
      orderTotal: number;
      /** Format: date */
      inHandDate?: string;
      orderNotes?: string;
      delivered: boolean;
      /** Format: date-time */
      deliveryTime?: string;
      processed: boolean;
      cancelled: boolean;
      billingAddress?: components['schemas']['Address'];
      shippingAddress?: components['schemas']['Address'];
      items: components['schemas']['OrderItem'][];
      paymentMethod?: string;
      creditCardLastFour?: string;
      creditCardType?: string;
      identityVerificationUrl?: string;
      gp?: number;
      showReviewPopup?: boolean;
      insurance?: components['schemas']['OrderInsurance'];
      declinedInsurance: boolean;
      rca?: boolean;
      promoCodeLastFour?: string;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}